Add Cesium Ion logo using vsgImGui
Added ImageComponent, which renders VSG texture data as an image in
ImGui.

Moved much (all?) of the UI code from csclient.cpp into a new UI
class which, in addition to creating the trackball etc., creates the
objects for rendering an ImGui interface.

A Cesium Ion logo is displayed in a fixed, transparent window. This
could be the start of a credit system display.