Remove broken traverse() code
The non-const Visitor version of TilesetNode::traverse() wasn't
checking if a tile's content is render content. So far, this function
is only used after a window resize event. It was crashing
mysteriously. Replaced the body of the function with a call to the
template used in the other traverse() functions.