Reuse curl handles in requests
It was a huge performance loss to create a new CURL handle for each
http request and then destroy it, especially with SSL, because libcurl
would make a new connection to the server for each request. if handles
are reused, then libcurl keeps persistent connections open. The time
for GET requests was reduced by 66%.