Do remove old state commands from tile StageGroup object
This was a bug left over from moving the BindDescriptorSet command for
view-dependent state to the world level. When rasters were attached,
the old state would be left in the state group, but this did not
affect rendering. It did cause very frustrating problems when trying
to write to tile data!