Change command line argument processing to be more useful
The worldviewer program now reads json files supplied as arguments and
determines the type of the resulting object, so "--world-file" is no
longer necessary. Additionally, URLs or file paths to 3D Tiles
tilesets can be specified directly, so there's no need to create a
world file or tileset file for quick tests.