Avoid copying Cesium Native assets and keep shared pointers to them
Instead of copying texture bytes into vsg::Array2D, point to it and
store a CesiumUtility::IntrusivePointer in the Array2D's auxilliary
data. This avoids keeping a redundant copy of the data -- which can be
significant in e.g. Google Photorealistic Tiles -- protects that data
from deletion, and causes it to be deleted when appropriate.